requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
async-timeout==5.0.1
attrs==25.4.0
cachetools==7.0.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.5
distro==1.9.0
discord.py==2.6.4
frozenlist==1.8.0
google-api-core==2.30.0
google-api-python-client==2.190.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.3.0
googleapis-common-protos==1.72.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
jiter==0.13.0
MarkupSafe==3.0.3
multidict==6.7.1
oauthlib==3.3.1
openai==2.24.0
packaging==26.0
pluggy==1.6.0
propcache==0.4.1
proto-plus==1.27.1
protobuf==6.33.5
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.2
pytest==9.0.2
python-dotenv==1.2.1
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
yarl==1.22.0
